rename unit.licences package to feat.licences
