+++ feature/licences/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/feat/licences/di/LicenceModule.kt
@@ -1,0 +1,8 @@
+package com.livefast.eattrash.raccoonforfriendica.feat.licences.di
+
+import org.koin.core.annotation.ComponentScan
+import org.koin.core.annotation.Module
+
+@Module
+@ComponentScan("com.livefast.eattrash.raccoonforfriendica.feat.licences")
+class LicenceModule
