--- feature/licences/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/unit/licences/di/LicenceModule.kt
@@ -1,8 +1,0 @@
-package com.livefast.eattrash.raccoonforfriendica.unit.licences.di
-
-import org.koin.core.annotation.ComponentScan
-import org.koin.core.annotation.Module
-
-@Module
-@ComponentScan("com.livefast.eattrash.raccoonforfriendica.unit.licences")
-class LicenceModule
